requirements.txt
keyboard==0.13.5
Pillow==11.2.1
pyautogui==0.9.54
pyperclip==1.9.0
pytesseract==0.3.13
